Implementirao sam prompt za brisanje pomocu react-confirm-alert

+++ src/pages/AppCars.js
@@ -1,5 +1,6 @@
 import carsService from "../services/CarsService";
 import { useContext, useEffect, useState } from "react";
+import { confirmAlert } from "react-confirm-alert";
 import { ListCars } from "../components/ListCars";
 
 export const AppCars = () => {
@@ -15,11 +16,25 @@ export const AppCars = () => {
   }, []);
 
   const handleDelete = (id) => {
-    const isDeleted = carsService.delete(id);
-    if (isDeleted) {
-      const newList = cars.filter((car) => car.id != id);
-      setCars(newList);
-    }
+    const confirmation = confirmAlert({
+      title: "Confirm Deletion",
+      message: "Are you sure you want to delete car?",
+      buttons: [
+        {
+          label: "Yes",
+          onClick: () => {
+            const isDeleted = carsService.delete(id);
+            if (isDeleted) {
+              const newList = cars.filter((car) => car.id != id);
+              setCars(newList);
+            }
+          },
+        },
+        {
+          label: "No",
+        },
+      ],
+    });
   };
 
   return (
